% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{GetHealthTopics}
\alias{GetHealthTopics}
\title{Requests all helath topics from API.
Calls enpoint http://atlas.ecdc.europa.eu/public/AtlasService/rest/GetHealthTopics}
\usage{
GetHealthTopics()
}
\description{
Requests all helath topics from API.
Calls enpoint http://atlas.ecdc.europa.eu/public/AtlasService/rest/GetHealthTopics
}

