% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{GetMeasuresResultsExportFile}
\alias{GetMeasuresResultsExportFile}
\title{Downloads measures as data frame from altas API at
http://atlas.ecdc.europa.eu/public/AtlasService/rest/GetMeasuresResultsExportFile}
\usage{
GetMeasuresResultsExportFile(healthTopicId)
}
\arguments{
\item{healthTopicId}{The healthTopicId to query data for}
}
\description{
Downloads measures as data frame from altas API at
http://atlas.ecdc.europa.eu/public/AtlasService/rest/GetMeasuresResultsExportFile
}

